dev/requirements-minimal.txt
# NIS Protocol Core Dependencies for Deployment
# This is a minimal version focusing on LLM providers and core functionality

# Core dependencies
redis>=6.0.0
hiredis>=2.0.0
pydantic>=1.9.0
fastapi>=0.95.0
uvicorn>=0.21.0
httpx>=0.23.0

# Memory system (lightweight)
numpy>=1.21.0
# sentence-transformers>=2.2.0  # Optional - heavy dependency
faiss-cpu>=1.7.0  # For vector search in memory

# LLM Integration (core functionality)
aiohttp>=3.8.0
tiktoken>=0.5.0
cachetools>=5.3.0
tenacity>=8.2.0  # For retrying API calls
async-timeout>=4.0.0

# Basic data processing
pandas>=1.3.0
requests>=2.26.0
websockets>=10.0
python-dotenv>=0.19.0
pyyaml>=5.4.1  # For configuration
jsonschema>=4.0.0

# Development and deployment
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# System utilities
psutil>=5.8.0  # For system monitoring ([health tracking](src/infrastructure/integration_coordinator.py))
protobuf>=3.17.3  # For protocol buffers
python-json-logger>=2.0.2  # For structured logging

# Optional heavy dependencies (install manually if needed):
# torch>=1.9.0  # For PyTorch-based features
# torchvision>=0.10.0  # For computer vision
# opencv-python>=4.6.0  # For image processing
# ultralytics>=8.0.0  # For YOLO object detection
# sentence-transformers>=2.2.0  # For embeddings
# transformers>=4.11.0  # For transformer models
# scipy>=1.7.0  # For scientific computing
# scikit-learn>=0.24.2  # For machine learning
# matplotlib>=3.4.0  # For plotting
# seaborn>=0.11.0  # For statistical plots
# sounddevice>=0.4.5  # For audio
# nltk>=3.6.0  # For NLP
# spacy>=3.1.0  # For comprehensive NLP 